+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * shipping_amount
   * ships_from_postal_code
 * Add support for transaction line items
+* Add support for tagged evidence in DisputeGateway#add_text_evidence (Beta release)
 
 ## 3.39.1
 * Fix spec to expect PayPal transactions to move to settling rather than settled